feat: loading section customizable height
