+++ packages/panels/src/Resources/RelationManagers/RelationManager.php
@@ -12,6 +12,7 @@ use Filament\Infolists\Infolist;
 use Filament\Pages\Page;
 use Filament\Resources\Concerns\InteractsWithRelationshipTable;
 use Filament\Resources\Pages\ViewRecord;
+use Filament\Support\Concerns\CanBeLazy;
 use Filament\Support\Enums\IconPosition;
 use Filament\Tables;
 use Filament\Tables\Actions\BulkAction;
@@ -27,6 +28,7 @@ use function Filament\authorize;
 class RelationManager extends Component implements Actions\Contracts\HasActions, Forms\Contracts\HasForms, Infolists\Contracts\HasInfolists, Tables\Contracts\HasTable
 {
     use Actions\Concerns\InteractsWithActions;
+    use CanBeLazy;
     use Forms\Concerns\InteractsWithForms;
     use Infolists\Concerns\InteractsWithInfolists;
     use InteractsWithRelationshipTable {
@@ -85,8 +87,6 @@ class RelationManager extends Component implements Actions\Contracts\HasActions,
     protected static ?string $badgeColor = null;
 
     protected static ?string $badgeTooltip = null;
-
-    protected static bool $isLazy = true;
 
     public function mount(): void
     {
@@ -519,24 +519,5 @@ class RelationManager extends Component implements Actions\Contracts\HasActions,
                 $this->getTableRecordUrlUsing(),
                 fn (Table $table, ?Closure $using) => $table->recordUrl($using),
             );
-    }
-
-    /**
-     * @return array<string, mixed>
-     */
-    public static function getDefaultProperties(): array
-    {
-        $properties = [];
-
-        if (static::isLazy()) {
-            $properties['lazy'] = true;
-        }
-
-        return $properties;
-    }
-
-    public static function isLazy(): bool
-    {
-        return static::$isLazy;
     }
 }
