+++ packages/widgets/src/Widget.php
@@ -2,14 +2,15 @@
 
 namespace Filament\Widgets;
 
+use Filament\Support\Concerns\CanBeLazy;
 use Illuminate\Contracts\View\View;
 use Livewire\Component;
 
 abstract class Widget extends Component
 {
-    protected static bool $isDiscovered = true;
+    use CanBeLazy;
 
-    protected static bool $isLazy = true;
+    protected static bool $isDiscovered = true;
 
     protected static ?int $sort = null;
 
@@ -67,11 +68,6 @@ abstract class Widget extends Component
         return static::$isDiscovered;
     }
 
-    public static function isLazy(): bool
-    {
-        return static::$isLazy;
-    }
-
     public function render(): View
     {
         return view(static::$view, $this->getViewData());
@@ -88,22 +84,11 @@ abstract class Widget extends Component
     /**
      * @return array<string, mixed>
      */
-    public static function getDefaultProperties(): array
+    public function getPlaceholderData(): array
     {
-        $properties = [];
-
-        if (static::isLazy()) {
-            $properties['lazy'] = true;
-        }
-
-        return $properties;
-    }
-
-    public function placeholder(): View
-    {
-        return view('filament::components.loading-section', [
+        return [
             'columnSpan' => $this->getColumnSpan(),
             'columnStart' => $this->getColumnStart(),
-        ]);
+        ];
     }
 }
